✅ Refactor test_enums to make them independent of previous imports
